feat: Add Hamburger component
This commit adds a new component called Hamburger. It represents a hamburger menu icon and includes SVG paths and attributes to define the shape and style of the icon. The component also includes JavaScript code to toggle the "expanded" class on the ".nav-links" element when the hamburger icon is clicked.