feat: Add slugify utility functions
This commit adds a new file called `slugify.ts` in the `utils` directory. The file contains two utility functions: `slugifyStr` and `slugifyAll`. `slugifyStr` takes a string as input and uses the `github-slugger` library to generate a slug. `slugifyAll` takes an array of strings and applies `slugifyStr` to each element, returning an array of slugs. These utility functions will be useful for generating slugs for blog post URLs and other purposes.