feat: Add new OG template for site
This commit adds a new file called `site.tsx` in the `utils/og-templates` directory. The file contains a template for generating Open Graph images specifically for the site. The template includes styling for the site title and description. This new template will enhance the visual representation of the site when shared on social media platforms.